requirements.txt
pynacl
requests
configparser
boto3